Add texture loading with sample

+++ Source/Engine/System/Misc/Vertex.cpp
@@ -43,6 +43,6 @@ void Shaft::Vertex::Init()
 		.add(bgfx::Attrib::Position, 3, bgfx::AttribType::Float)
 		.add(bgfx::Attrib::Normal, 4, bgfx::AttribType::Uint8, true, true)
 		.add(bgfx::Attrib::Tangent, 4, bgfx::AttribType::Uint8, true, true)
-		.add(bgfx::Attrib::TexCoord0, 2, bgfx::AttribType::Int16, true, true)
+		.add(bgfx::Attrib::TexCoord0, 2, bgfx::AttribType::Float, true, true)
 		.end();
 }
